refactor: move loading modal to overlay component
