+++ src/components/Layout/NetworkSelect/NetworkSelect.tsx
@@ -10,7 +10,7 @@ import React, { FunctionComponent, useContext } from "react";
 import { ChainId, ChainInfoObject } from "../../../constants/chain-info";
 import { useProviderContext } from "../../../common/contexts/provider";
 import { getChainInfo } from "../../../common/utils/chain-utils";
-import { LoadingModal } from "../../Demo/LoadingModal";
+import { LoadingModal } from "../../UI/Overlay";
 
 interface NetworkSelectViewProps {
   onChange: (network: ChainInfoObject) => void;
