+++ src/components/UI/Overlay/index.tsx
@@ -1,1 +1,2 @@
 export * from "./StaticOverlay";
+export * from "./LoadingModal";
